feat: add actions to teams tables
