+++ app/components/shared/Dropdown/Dropdown.tsx
@@ -1,3 +1,4 @@
+import { IconCaretDown } from "@pokt-foundation/pocket-blocks"
 import * as DropdownMenu from "@radix-ui/react-dropdown-menu"
 import clsx from "clsx"
 import styles from "./styles.css"
@@ -35,5 +36,40 @@ export const Dropdown: React.FC<DropdownProps> = ({
   )
 }
 
+type DropdownTriggerProps = {
+  label: string
+}
+
+type DropdownItemProps = {
+  label: string
+  variant?: "default" | "green"
+  action: () => void
+}
+
+function DropdownTrigger({ label }: DropdownTriggerProps) {
+  return (
+    <div className="dropdown-trigger">
+      <div className="dropdown-trigger__label">{label}</div>
+      <div className="dropdown-trigger__arrow">
+        <IconCaretDown className="pokt-icon" height="12px" width="12px" />
+      </div>
+    </div>
+  )
+}
+
+function DropdownItem({ label, action, variant }: DropdownItemProps) {
+  return (
+    <button
+      className={clsx({
+        "dropdown-item": true,
+        "dropdown-item--green": variant === "green",
+      })}
+      onClick={action}
+    >
+      {label}
+    </button>
+  )
+}
+
 export default Dropdown
-export { DropdownMenu }
+export { DropdownMenu, DropdownItem, DropdownTrigger }
