+++ app/components/shared/Dropdown/styles.css
@@ -49,3 +49,66 @@
   font-size: 18px;
   margin: 0.5em 0;
 }
+
+.dropdown-trigger {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.dropdown-trigger__arrow {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-left: 8px;
+  padding: 6px;
+  background: #18222c;
+  border: 1px solid #222e3b;
+  border-radius: 2px;
+}
+
+.dropdown-item {
+  all: unset;
+  display: block;
+  box-sizing: border-box;
+  padding: 8px 12px;
+  min-width: 152px;
+  border-radius: 2px;
+  color: var(--color-white-main);
+  background-color: #293746;
+  cursor: pointer;
+}
+
+.dropdown-item:not(:first-child) {
+  margin-top: 4px;
+}
+
+.dropdown-item:hover:not(:first-child) {
+  margin-top: 4px;
+}
+
+.dropdown-item:hover {
+  all: unset;
+  display: block;
+  box-sizing: border-box;
+  padding: 8px 12px;
+  min-width: 152px;
+  border-radius: 2px;
+  color: var(--color-white-main);
+  background-color: #29374680;
+  cursor: pointer;
+}
+
+.dropdown-item--green {
+  background-color: var(--color-primary-main);
+  color: var(--color-black-main);
+}
+
+.dropdown-item--green:hover {
+  background-color: var(--color-primary-light);
+  color: var(--color-black-main);
+}
+
+.dropdown-teams__content {
+  padding: 4px;
+}
